pyproject.toml
[project]
name = "chat-response-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.124.4",
    "openai>=2.11.0",
    "pillow>=12.0.0",
    "playwright>=1.57.0",
    "pydantic>=2.12.5",
]
